# docs/swagger.yaml
components:
  schemas:
...
    Event:
      properties:
        application_end_time:
          format: date-time
          type: string
        application_start_time:
          format: date-time
          type: string
        event_end_time:
          format: date-time
          type: string
        event_start_time:
          format: date-time
          type: string
        game_version:
          $ref: '#/components/schemas/GameVersion'
        id:
          type: integer
        is_current:
          type: boolean
        is_locked:
          type: boolean
        is_main_event:
          type: boolean
        is_public:
          type: boolean
        max_size:
          type: integer
        name:
          type: string
        patch:
          type: string
        teams:
          items:
            $ref: '#/components/schemas/Team'
          type: array
          uniqueItems: false
        uses_medals:
          type: boolean
        waitlist_size:
          type: integer
      required:
      - application_end_time
      - application_start_time
      - event_end_time
      - event_start_time
      - game_version
      - id
      - is_current
      - is_locked
      - is_main_event
      - is_public
      - max_size
      - name
      - teams
      - uses_medals
      - waitlist_size
      type: object
...
    ExtendedSignup:
      properties:
        expected_playtime:
          type: integer
        extra:
          type: string
        highest_character_levels:
          additionalProperties:
            type: integer
          type: object
        needs_help:
          type: boolean
        partner:
          $ref: '#/components/schemas/NonSensitiveUser'
        partner_id:
          type: integer
        partnerWish:
          type: string
        playtimes_in_last_events_per_day_in_hours:
          additionalProperties:
            type: number
          type: object
        team_id:
          type: integer
        team_lead:
          type: boolean
        timestamp:
          format: date-time
          type: string
        user:
          $ref: '#/components/schemas/NonSensitiveUser'
        wants_to_help:
          type: boolean
      required:
      - expected_playtime
      - highest_character_levels
      - playtimes_in_last_events_per_day_in_hours
      - team_lead
      - timestamp
      - user
      type: object
...
    NonSensitiveUser:
      properties:
        account_name:
          type: string
        discord_id:
          type: string
        discord_name:
          type: string
        display_name:
          type: string
        id:
          type: integer
        twitch_id:
          type: string
        twitch_name:
          type: string
      required:
      - display_name
      - id
      type: object
...
    Signup:
      properties:
        expected_playtime:
          type: integer
        extra:
          type: string
        needs_help:
          type: boolean
        partner:
          $ref: '#/components/schemas/NonSensitiveUser'
        partner_id:
          type: integer
        partnerWish:
          type: string
        team_id:
          type: integer
        team_lead:
          type: boolean
        timestamp:
          format: date-time
          type: string
        user:
          $ref: '#/components/schemas/NonSensitiveUser'
        wants_to_help:
          type: boolean
      required:
      - expected_playtime
      - team_lead
      - timestamp
      - user
      type: object
    SignupCreate:
      properties:
        expected_playtime:
          type: integer
        extra:
          type: string
        needs_help:
          type: boolean
        partner_account_name:
          type: string
        wants_to_help:
          type: boolean
      required:
      - expected_playtime
      type: object
...
    Team:
      properties:
        abbreviation:
          type: string
        allowed_classes:
          items:
            type: string
          type: array
          uniqueItems: false
        color:
          type: string
        discord_role_id:
          type: string
        event_id:
          type: integer
        id:
          type: integer
        name:
          type: string
      required:
      - allowed_classes
      - event_id
      - id
      - name
      type: object
...
    GameVersion:
      enum:
      - poe1
      - poe2
      type: string
      x-enum-varnames:
      - PoE1
      - PoE2
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: Liberatorist@gmail.com
    name: Liberator
  description: This is the backend API for the BPL project.
  title: BPL Backend API
  version: "2.0"
openapi: 3.1.0
paths:
...
  /events/{event_id}:
    delete:
      description: Deletes an event
      operationId: DeleteEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - event
    get:
      description: Fetches an event by id
      operationId: GetEvent
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - event
...
  /events/{event_id}/signups:
    get:
      description: Fetches all signups for the event
      operationId: GetEventSignups
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ExtendedSignup'
                type: array
          description: OK
      security:
      - BearerAuth: []
      tags:
      - signup
  /events/{event_id}/signups/{user_id}:
    delete:
      description: Deletes a user's signup for the event
      operationId: DeleteSignup
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      - description: User Id
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      tags:
      - signup
  /events/{event_id}/signups/self:
    get:
      description: Fetches an authenticated user's signup for the event
      operationId: GetPersonalSignup
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Signup'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - signup
    put:
      description: Creates a signup for the authenticated user
      operationId: CreateSignup
      parameters:
      - description: Event Id
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/SignupCreate'
                description: Signup
                summary: signupCreate
        description: Signup
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Signup'
          description: Created
      security:
      - BearerAuth: []
      tags:
      - signup
...